+++ HISTORY.rst
@@ -4,5 +4,7 @@ History
 
 unreleased
 ------------------
-
 * First release on PyPI.
+* Adeed cli interface
+* Added utils tools
+* Added api manga